gstreamer: convert from libcamera colorspace to GStreamer colorimetry.
Libcamera StreamConfiguration class has colorSpace attribute, which
holds the colorspace that is being applied to the camera after the
validation of the camera configuration.

Add a std::map that will map libcamera colorspace to GStreamer colorimetry and
find the colorimetry corresponding to the colorspace that is being applied to
the camera.If the colorimetry is found, the helper function
libcamera_colorimetry_to_gst_string() will take in a std::string representation
of colorimetry and convert it into G_TYPE_STRING so that it can be pushed into the
caps. No need to check if the colorimetry if found is valid as we are working with
known predefined colorimetry.

Signed-off-by: Rishikesh Donadkar <rishikeshdonadkar@gmail.com>