gstreamer: convert from libcamera colorspace toGStreamer colorimetry.
Libcamera StreamConfiguration class has colorSpace attribute, which
holds the colorspace that is being applied to the camera after the
validation of the camera configuration.

Map the libcamera colorspace to GStreamer colorimetry and find
the colorimetry corresponding to the colorspace that is being applied to
the camera. This colorimetry if found will be pushed in the caps.

Signed-off-by: Rishikesh Donadkar <rishikeshdonadkar@gmail.com>