gstreamer: convert from libcamera colorspace to GStreamer colorimetry.
Libcamera StreamConfiguration class has colorSpace attribute, which
holds the colorspace that is being applied to the camera after the
validation of the camera configuration.

Add a std::map that will map libcamera colorspace to GStreamer colorimetry and
find the colorimetry corresponding to the colorspace that is being applied to
the camera.If the colorimetry is found, the helper function
libcamera_colorimetry_to_gst_string() will take in a std::string representation
of colorimetry and convert it into G_TYPE_STRING so that it can be pushed into the
caps. If the colorimetry is not from the standard ones, create the
colorimetry instance mapping the fields of colorspace with the fields
of the colorimetry. This colorimetry will be pushed in the caps and the
downstream will decide if the colorimetry is acceptable.

Signed-off-by: Rishikesh Donadkar <rishikeshdonadkar@gmail.com>