gstreamer: Update the obtained colorimetry in caps.
If the colorspace is set in the StreamConfiguration, convert the
libcamera colorspace to GStreamer colorimetry using the function
colorimetry_from_colorspace() which will take in the ColorSpace value
and return the gst string representation of the colorimetry.

If the colorimetry corresponding to the colorspace set in the
StreamConfiguration is not available in GStreamer remove the
field colorimetry from the caps structure if present.

Signed-off-by: Rishikesh Donadkar <rishikeshdonadkar@gmail.com>