gstreamer: Convert from GStreamer colorimetry to libcamera ColorSpace
Add map ColorimetryToColorSpace that maps the GStreamer colorimetry to
libcamera ColorSpace and the function colorspace_from_colorimetry()
to convert from gstreamer colorimetry to libcamera colorspace using the map.
The obtained colorspace form the conversion will be applied to the
colorSpace field of the StreamConfiguration. If one of the field of the
colorimetry is not available then the default field is applied to the
colorspace corresponing to that colorimetry. The default colorimetry
is obtained from the GstVideoInfo.

Signed-off-by: Rishikesh Donadkar <rishikeshdonadkar@gmail.com>