gstreamer: Add multiple colorimetry support
This patch aims to add support for multiple colorimetry.
retrive the structure with best suitable resolution, add it
into a fresh new caps and normalize the caps using
gst_caps_normalize(). This will break the structrue into as many
structures as the number of colorimetry specified in the gstreamer
pipeline.

Iterate over each structure, retrieve the colorimetry string, convert to
colorspace and validate the camera configuration. Retrieve the colorspace
after validation, convert to colorimetry and check if the it is same as
the coloriemtry requested.

If none of the colorimetry requested is supported by the camera (ie not
the same after validation) then set the colorspace to the value that was
present before trying new colorimetries.

--------------------
1. get the index of the structure with best sutaible resolution from
   gst_libcamera_configure_stream_from_caps() by changing its return
   type.
2. Make a fresh caps instance and add the structure present at the
   index into the new caps using gst_caps_copy_nth() and normalize the
   caps.
3. Itreate over each structure in the new caps.
   in each iteration
   1) get the coloriemtry form structure
   2) convert to colorspace
   3) validate stream cfg
   4) get the colorspace form the stream cfg irrespective of weather it
      is changed.
   5) Convert the colorspace to colorimetry, check if the coloriemtry is
      the same as the one in (1). If same then break the
      loop(colorimetry can pass negotiation.), if different then set the
      colorspace in stream configuration to the colorspace that was
      present before (1)
4. The appropriate colorspace will be set in the stream configuration as
   a result of the above loop. Again validate, convert colorspace to
   colorimetry and push it into the caps.
--------------------