gstreamer: Add multiple colorimetry support
This patch aims to add support for multiple colorimetry.

Append copy of the structure with best suitable resolution into a
fresh new caps and normalize the caps using gst_caps_normalize().
This will return caps where the colorimetry list is expanded.
The ncaps will contain as many structures as the number of
colorimetry specified in the gstreamer pipeline.

Iterate over each structure in the ncaps, retrieve the colorimetry string,
convert to colorspace and validate the camera configuration.
Retrieve the colorspace after validation, convert to colorimetry and check
if the it is same as the colorimetry requested.

If none of the colorimetry requested is supported by the camera (ie not
the same after validation) then set the stream_cfg to the previous configuration
that was present before trying new colorimetry.