gstreamer: Add multiple colorimetry support
This patch aims to add support for multiple colorimetry.
Append copy of the structure with best suitable resolution into a
fresh new caps and normalize the caps using gst_caps_normalize().
This will retrun a caps where the coloriemtry list is expanded.
The ncaps will contain as many structures as the number of
colorimetry specified in the gstreamer pipeline.

Iterate over each structure, retrieve the colorimetry string, convert to
colorspace and validate the camera configuration. Retrieve the colorspace
after validation, convert to colorimetry and check if the it is same as
the coloriemtry requested.

If none of the colorimetry requested is supported by the camera (ie not
the same after validation) then set the colorspace to the value that was
present before trying new colorimetries.