gstreamer: Provide mulitple colorimetry support
Iterate over each GValue in the colorimetry list,
retrieve the colorimetry string, convert to colorspace using the
function colorspace_from_colorimetry() and validate the camera configuration.
Retrieve the colorspace after validation, convert to colorimetry and
check if the it is same as the colorimetry requested.

If none of the colorimetry requested is supported by the camera (i.e. not
the same after validation) then set the stream_cfg to the previous
configuration that was present before trying new colorimetry.

Signed-off-by: Rishikesh Donadkar <rishikeshdonadkar@gmail.com>
Signed-off-by: Umang Jain <umang.jain@ideasonboard.com>