gstreamer: Provide mulitple colorimetry support
Append copy of the structure with best suitable resolution into a
fresh new caps and normalize the caps using gst_caps_normalize().
This will return caps where the colorimetry list is expanded.
The ncaps will contain as many structures as the number of
colorimetry specified in the GStreamer pipeline.

Iterate over each structure in the ncaps, retrieve the colorimetry string,
convert to colorspace using the function colorspace_from_colorimetry(),
apply the resultant colorspace, and validate the camera configuration.
Retrieve the colorspace after validation, check if it is same as the
colorspace before validation.

If none of the colorspace requested is supported by the camera (i.e. not
the same after validation) then set the stream_cfg to the previous
configuration that was present before trying new colorimetry.

Signed-off-by: Rishikesh Donadkar <rishikeshdonadkar@gmail.com>
Signed-off-by: Umang Jain <umang.jain@ideasonboard.com>