gstreamer: Provide framerate support for libcamerasrc.
Add a field of the type ControlList to GstLibcameraSrcState.

Get the framerate from the caps and convert it to FrameDuration.
Set the FrameDurationLimits control in the initControls_

Passing initControls_ at camera::start() doesn't tell us if the controls
have been applied to the camera successfully or not (or they have adjusted in
some fashion). Until that is introduced in camera::start() API,
we need to carry out such handling on the application side.

Check the frame_duration whether it is in-bound to the max / min
frame-duration supported by the camera using the function and also check
if it is capable to pass the negotiation.
gst_libcamera_bind_framerate().

If the frame_duartion is out of bounds, set the framerate to 0/1 in
container which will be later exposed to gstreamer.

Pass the initControls_ at the time of starting camera.

Solve the complications in exposing the correct framerate due to loss in
precision as a result of casting the frameduration to int64_t(which is
required in libcamera to set the FrameDurationLimits control).

Example -

* Suppose the framerate requested is 35/1. The framerate is read form
  the caps in the from of fraction that has a numerator and
  denominator.

* Converting to FrameDuration (in microseconds)
    (1 * 10^6) / 35 = 28571.4286
    int64_t frame_duration = 28571
    (the precision here is lost.)
* To expose the framerate in caps, Inverting the frame_duration to get
  back the framerate and converting to Seconds.
    double framerate = 10^6 / 28571
    and
    10^6/28571 which is close but not equal to 35/1 will fail the negotiation.

To solve the above problem, Store the framerate requested through
negotiated caps in the GValue of the type GST_TYPE_FRACTION.

Try to apply the framerate and check if the floor value of framerate that gets applied
closely matches with the floor value framerate requested in the negotiated caps. If
the value matches expose the framerate that is stored in the container, else expose
the framerate in the new caps as 0/1.

Pass the initControls_ at camera->start().

Signed-off-by: Rishikesh Donadkar <rishikeshdonadkar@gmail.com>