gstreamer: Provide framerate support for libcamerasrc.
Add a field of the type ControlList to GstLibcameraSrcState.

Get the framerate from the caps and convert it to FrameDuration.
Set the FrameDurationLimits control in the initControls_

Passing initCtrls_ at camera::start() doesn't tell us if the controls
have been applied to the camera successfully or not (or they have adjusted in
some fashion). Until that is introduced in camera::start() API,
we need to carry out such handling on the application side.
Check the frame_duration whether it is in-bound to the max / min
frame-duration supported by the camera.
If the framer_duartion is out of bounds, bind the frame_duration
between the max and min FrameDuration that is supported by the camera.

Pass the initControls_ at the time of starting camera.

Change in the existing code structure[RFC]: Shift the call to
camera::cofigure() above the for loop which does not touch
anything form libcamera just reads from it. This is done because
the bounds checking for frame duration cannot be done without the
call to camera::configure(). This will ease exposing the  framerate to
GStreamer. The framerate can be exposed with the same new CAPS event
which is exposing the width, height and the colorimetry.

Solve the complications in exposing the correct framerate due to loss in
precision as a result of casting the frameduration to int64_t(which is
required in libcamera to set the FrameDurationLimits control).

Example -

* Suppose the framerate requested is 35/1. The framerate is read form
  the caps in the from of fraction that has a numerator and
  denominator.

* Converting to FrameDuration (in microseconds)
    (1 * 10^6) / 35 = 28571.4286
    int64_t frame_duration = 28571
    (the precision here is lost.)
* To expose the framerate in caps, Inverting the frame_duration to get
  back the framerate and converting to Seconds.
    double framerate = 10^6 / 28571
    and
    10^6/28571 which is close but not equal to 35/1 will fail the negotiation.

To solve the above problem, introduce a global std::pair<int, int> variable that
stores the numerator and denominator of the framerate in the negotiatied caps.
Try to apply the framerate and check if the floor value of framerate that gets applied
closely matches with the floor value framerate requested in the negotiatied caps. If
the value matches set the framerate that is stored in the global container, else set the
framerate to 0/1.

Signed-off-by: Umang Jain <umang.jain@ideasonboard.com>
Signed-off-by: Rishikesh Donadkar <rishikeshdonadkar@gmail.com>