gstreamer: Provide framerate support for libcamerasrc.
Control the framerate by setting the controls::FrameDurationLimits
in an ControlList and passing the control list at camera::start(). This
requires converting the framerate which of the type
GST_TYPE_FRACTION into FrameDuration which of the type int64_t. This
conversion causes loss of precision and the precise framerate cannot be
obtained by Inverting the control::FrameDurationLimits value.

Example -

* Suppose the framerate requested is 35/1. The framerate is read form
  the caps in the from of fraction that has a numerator and
  denominator.

* Converting to FrameDuration (in microseconds)
    (1 * 10^6) / 35 = 28571.4286
    int64_t frame_duration = 28571
    (the precision here is lost.)
* To expose the framerate in caps, Inverting the frame_duration to get
  back the framerate and converting to Seconds.
    double framerate = 10^6 / 28571
    and
    10^6/28571 which is close but not equal to 35/1 will fail the negotiation.

Get the framerate from the caps and convert it to FrameDuration.

Clamp the frame_duration between the min/max FrameDuration supported
by the camera. Set the FrameDurationLimits control in the initControls_.
Store the clamped/unclamped framerate in the container to be retrieved later.

Set the bound checked framerate from the container into the caps.

Pass the ControlList initControls_ at camera->start().

Signed-off-by: Rishikesh Donadkar <rishikeshdonadkar@gmail.com>