gstreamer: Configure the camera before exposing the caps.
In libcamerasrc capabilities are exposed on the src pad before
configuring the camera. To add support to control and
negotiate the framerate, the controls::FrameDuration needs to be
bound checked between the min/max values that camera can support
and later the applied framerate needs to be exposed along with
resolutions and colorimetry through caps. Valid bounds of the
controls::FrameDuration cannot be known before the
configuration of the camera.

Shift the camera::configure() before the for loop that is exposing
resolutions, colorimetry to the GStreamer through a new CAPS event.
Through this we can know the valid bounds of the FrameDuration and
clam the frame_duration accordingly before applying to the camera.
Through this caps can be exposed without a need of additional new
CAPS event for framerate.

Signed-off-by: Rishikesh Donadkar <rishikeshdonadkar@gmail.com>