gstreamer: Provide mulitple colorimetry support for libcamerasrc
Iterate over each GValue in the colorimetry list,
retrieve the colorimetry string, convert to colorspace,
set the colorspace in the  stream_cfg and validate the camera configuration.
Retrieve the colorspace after validation, convert to colorimetry and
check if the it is same as the colorimetry requested.

If none of the colorimetry requested is supported by the camera (i.e. not
the same after validation) then set the stream_cfg to the previous
configuration that was present before trying new colorimetry.